*: Port to new Azure SDK
Most of the changes are because pageblob.Client represents a specific
blob in storage, so it can be passed instead of a triplet of general
blob storage client, container name and blob name.

On the authentication side, we switch to the azidentity package, but
we preserve the possibility of using the discouraged shared key
crendetials if --stgaccountkey parameter is passed. Otherwise we will
use the managed identity, with an additional requirement that the
managed account already has a role that allows the account to save new
blobs.